feat: custom exceptions, extended ruff rules, shared test fixtures, pre-commit hooks (#37, #42, #43, #44)
- Add optimizer/exceptions.py with OptimizerError hierarchy (ConfigurationError,
  DataError, ConvergenceError, ValidationError, OptimizationError) replacing all
  bare ValueError/TypeError/RuntimeError raises across 15 source files
- Extend ruff lint rules (B, SIM, S, RUF, C4, PTH), fix all violations, remove
  black from dev deps, globally ignore RUF002/RUF003 for scientific unicode
- Create shared test fixtures in tests/conftest.py and tests/optimization/conftest.py,
  deduplicate returns_df and make_hmm_result across 7 test files
- Add .pre-commit-config.yaml with trailing-whitespace, end-of-file-fixer,
  check-yaml, check-toml, check-added-large-files, ruff, ruff-format hooks
- Add pre-commit>=3.5 to dev dependencies
- Apply cosmetic whitespace/EOF fixes from pre-commit across api/, cli/, frontend/

+++ optimizer/factors/_selection.py
@@ -138,16 +138,15 @@ def apply_sector_balance(
     result = list(selected)
 
     for sector, target_w in target_weights.items():
-        min_n = max(0, int(round((target_w - tolerance) * n_selected)))
-        max_n = int(round((target_w + tolerance) * n_selected))
+        min_n = max(0, round((target_w - tolerance) * n_selected))
+        max_n = round((target_w + tolerance) * n_selected)
 
         current_n = (selected_sectors == sector).sum()
 
         if current_n < min_n:
             # Under-represented: add highest-scoring non-selected from this sector
             candidates = sector_labels.index[
-                (sector_labels == sector)
-                & (~sector_labels.index.isin(selected))
+                (sector_labels == sector) & (~sector_labels.index.isin(selected))
             ]
             candidate_scores = (
                 scores.reindex(candidates).dropna().sort_values(ascending=False)
